requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-Bcrypt>=1.0.1
python-dotenv>=1.0.0
# Phase 2 — file upload and Excel parsing
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
# Phase 3 — forecasting models
statsmodels>=0.14.0
scikit-learn>=1.3.0
# Testing
pytest>=8.0.0
